refactor: rename utils.ts to index.ts to eliminate redundancy
## Summary
- Renamed src/utils/utils.ts to src/utils/index.ts to remove stuttering
- Updated all import paths throughout codebase to reference new location
- Improved code clarity and follows conventional index.ts naming pattern

## Changes
- File rename: src/utils/utils.ts → src/utils/index.ts
- Import path updates: utils/utils.js → utils/index.js
- Updated 16 files across src/ and tests/ directories
- All tests continue to pass with new import paths

## Benefits
- Eliminates redundant naming (utils/utils)
- Follows Node.js module conventions (index.js as entry point)
- Improves developer experience and code readability
